Fix bug when update user

+++ pkg/controllers/blog.go
@@ -3,11 +3,13 @@ package controllers
 import (
 	"Blog_API/pkg/domain"
 	"Blog_API/pkg/models"
-	"github.com/labstack/echo/v4"
+	"Blog_API/pkg/types"
 	"net/http"
 	"strconv"
+	"time"
+	"github.com/labstack/echo/v4"
+	"gorm.io/gorm"
 )
-
 
 // Parent struct to implement interface binding
 type blogController struct {
@@ -23,14 +25,28 @@ func NewBlogController(svc domain.BlogService) domain.BlogController {
 
 // CreateBlogPost implements domain.BlogController.
 func (ctr *blogController) CreateBlogPost(c echo.Context) error {
-	blogPost := &models.BlogPost{}
-	if err := c.Bind(blogPost); err != nil {
+	reqBlogPost := &types.BlogPostRequest{}
+	if err := c.Bind(reqBlogPost); err != nil {
 		return c.JSON(http.StatusBadRequest, "Invalid data request")
 	}
-	if err := ctr.svc.CreateBlogPost(blogPost); err != nil {
+	if err := reqBlogPost.Validate(); err != nil {
 		return c.JSON(http.StatusBadRequest, err.Error())
 	}
-	return c.JSON(http.StatusOK, blogPost)
+	currentTime := time.Now()
+	blog := &models.BlogPost{
+		UserID:       reqBlogPost.UserID,
+		Title:        reqBlogPost.Title,
+		ContentText:  reqBlogPost.ContentText,
+		PhotoURL:     reqBlogPost.PhotoURL,
+		Description:  reqBlogPost.Description,
+		Category:     reqBlogPost.Category,
+		IsPublished:  reqBlogPost.IsPublished,
+		PublishedAt:  &currentTime,
+	}
+	if err := ctr.svc.CreateBlogPost(blog); err != nil {
+		return c.JSON(http.StatusBadRequest, err.Error())
+	}
+	return c.JSON(http.StatusCreated, blog)
 }
 
 // GetBlogPost implements domain.BlogController.
@@ -61,14 +77,34 @@ func (ctr *blogController) GetBlogPosts(c echo.Context) error {
 
 // UpdateBlogPost implements domain.BlogController.
 func (ctr *blogController) UpdateBlogPost(c echo.Context) error {
-	blogPost := &models.BlogPost{}
+	blogPost := &types.UpdateBlogPostRequest{}
 	if err := c.Bind(blogPost); err != nil {
 		return c.JSON(http.StatusBadRequest, "Invalid data request")
 	}
-	if err := ctr.svc.UpdateBlogPost(blogPost); err != nil {
+	tempID := c.Param("id")
+	id, err := strconv.ParseUint(tempID, 10, 64)
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, "Invalid data request")
+	}
+	existingBlogPost, err := ctr.svc.GetBlogPost(uint(id))
+	if err != nil {
 		return c.JSON(http.StatusBadRequest, err.Error())
 	}
-	return c.JSON(http.StatusOK, blogPost)
+	if existingBlogPost.ID == 0 {
+		return c.JSON(http.StatusNotFound, "Blog post not found")
+	}
+	blog := &models.BlogPost{
+		Model: gorm.Model{ID: uint(existingBlogPost.ID), CreatedAt: existingBlogPost.CreatedAt, UpdatedAt: time.Now(), DeletedAt: existingBlogPost.DeletedAt},
+		Title:        blogPost.Title,
+		ContentText:  blogPost.ContentText,
+		PhotoURL:     blogPost.PhotoURL,
+		Description:  blogPost.Description,
+		Category:     blogPost.Category,
+	}
+	if err := ctr.svc.UpdateBlogPost(blog); err != nil {
+		return c.JSON(http.StatusBadRequest, err.Error())
+	}
+	return c.JSON(http.StatusOK, "Blog post updated successfully")
 }
 
 // DeleteBlogPost implements domain.BlogController.
@@ -76,6 +112,13 @@ func (ctr *blogController) DeleteBlogPost(c echo.Context) error {
 	id, err := strconv.ParseUint(c.Param("id"), 10, 64)
 	if err != nil {
 		return c.JSON(http.StatusBadRequest, "Invalid data request")
+	}
+	existingBlogPost, err := ctr.svc.GetBlogPost(uint(id))
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, err.Error())
+	}
+	if existingBlogPost.ID == 0 {
+		return c.JSON(http.StatusNotFound, "Blog post not found")
 	}
 	if err := ctr.svc.DeleteBlogPost(uint(id)); err != nil {
 		return c.JSON(http.StatusBadRequest, err.Error())
